service-livraisons/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
strawberry-graphql[fastapi]==0.227.0
pydantic==2.6.4
pymongo==4.7.2
python-multipart==0.0.9
